ecommerce/requirements.txt
python-dotenv
pydantic
pydantic_settings
azure-cosmos
fastapi
uvicorn
pydantic
